word_segmentation/requirements.txt
wordcutpy == 0.1.2
pylcs == 0.0.6